More from PR #67

+++ src/main/java/org/jenkinsci/plugins/vsphere/tools/CloudProvisioningRecord.java
@@ -123,7 +123,7 @@ public final class CloudProvisioningRecord {
 
     private double calcFullness() {
         final int maxToProvision = calcMaxToProvision();
-        return ((double) calcCurrentCommitment()) / (double) maxToProvision;
+        return calcCurrentCommitment() / (double) maxToProvision;
     }
 
     boolean hasCapacityForMore() {
